Fixes, add form validation

+++ OnboardingTask.Server/Controllers/CustomersController.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Data.SqlClient;
 using Microsoft.EntityFrameworkCore;
 using OnboardingTask.Server.Models;
 
@@ -22,43 +23,66 @@ namespace OnboardingTask.Server.Controllers
 		}
 
 		[HttpGet]
-		public async Task<ActionResult<IEnumerable<Customer>>> GetCustomers()
+		public async Task<ActionResult<IEnumerable<Customer>>?> GetCustomers()
 		{
-			return await _context.Customers.ToListAsync();
+			try
+			{
+				return await _context.Customers.ToListAsync();
+			}
+			catch (Exception)
+			{
+				return null;
+			}
 		}
 
 		[HttpGet("{id}")]
-		public async Task<ActionResult<Customer>> GetCustomer(int id)
+		public async Task<ActionResult<Customer>?> GetCustomer(int id)
 		{
-			var customer = await _context.Customers.FindAsync(id);
-
-			if (customer == null)
+			if (id == 0)
 			{
-				return NotFound();
+				return BadRequest(new { Message = "Invalid customer ID" });
 			}
-
-			return customer;
-		}
-
-		[HttpPut("{id}")]
-		public async Task<IActionResult> PutCustomer(int id, Customer customer)
-		{
-			if (id != customer.Id)
-			{
-				return BadRequest();
-			}
-
-			_context.Entry(customer).State = EntityState.Modified;
 
 			try
 			{
+				var customer = await _context.Customers.FindAsync(id);
+
+				if (customer == null)
+				{
+					return NotFound(new { Message = "Customer not found." });
+				}
+
+				return customer;
+			}
+			catch (Exception)
+			{
+				return NotFound(new { Message = "Customer not found." });
+			}
+		}
+
+		[HttpPut("{id}")]
+		public async Task<ActionResult<Customer>> PutCustomer(int id, Customer customer)
+		{
+			if (!ModelState.IsValid)
+			{
+				return BadRequest(ModelState);
+			}
+
+			if (id != customer.Id)
+			{
+				return BadRequest(new { Message = "Customer ID does not match." });
+			}
+
+			try
+			{
+				_context.Entry(customer).State = EntityState.Modified;
 				await _context.SaveChangesAsync();
 			}
 			catch (DbUpdateConcurrencyException)
 			{
 				if (!CustomerExists(id))
 				{
-					return NotFound();
+					return NotFound(new { Message = "Customer not found." });
 				}
 				else
 				{
@@ -66,31 +90,60 @@ namespace OnboardingTask.Server.Controllers
 				}
 			}
 
-			return NoContent();
+			return customer;
 		}
 
 		[HttpPost]
 		public async Task<ActionResult<Customer>> PostCustomer(Customer customer)
 		{
-			_context.Customers.Add(customer);
-			await _context.SaveChangesAsync();
+			if (!ModelState.IsValid)
+			{
+				return BadRequest(ModelState);
+			}
 
-			return CreatedAtAction("GetCustomer", new { id = customer.Id }, customer);
+			try
+			{
+				_context.Customers.Add(customer);
+				await _context.SaveChangesAsync();
+
+				return CreatedAtAction("GetCustomer", new { id = customer.Id }, customer);
+			}
+			catch
+			{
+				return BadRequest(new { Message = "An error occured please try again." });
+			}
 		}
 
 		[HttpDelete("{id}")]
 		public async Task<IActionResult> DeleteCustomer(int id)
 		{
-			var customer = await _context.Customers.FindAsync(id);
-			if (customer == null)
+			if (id == 0)
 			{
-				return NotFound();
+				return BadRequest(new { Message = "Invalid customer ID" });
 			}
 
-			_context.Customers.Remove(customer);
-			await _context.SaveChangesAsync();
+			try
+			{
+				var customer = await _context.Customers.FindAsync(id);
+				if (customer == null)
+				{
+					return NotFound(new { Message = "Customer not found." });
+				}
 
-			return NoContent();
+				_context.Customers.Remove(customer);
+				await _context.SaveChangesAsync();
+
+				return Ok(new { Message = "Customer deleted successfully." });
+			}
+
+			catch (DbUpdateException e)
+			{
+				return BadRequest(new { Error = true, Message = (e.InnerException is SqlException es && es.Number == 547) ? "Cannot delete record in use." : "An error occurred try again" });
+			}
+			catch (Exception)
+			{
+				return BadRequest(new { Message = "An error occurred try again." });
+			}
 		}
 
 		private bool CustomerExists(int id)
